+++ src/Myxomatosis/Configuration/DefaultConfiguration.cs
@@ -33,7 +33,7 @@ namespace Myxomatosis.Configuration
 
         public IRabbitMqClientLogger Logger { get; internal set; }
 
-        public int PrefetchCount { get; internal set; }
+        public ushort PrefetchCount { get; internal set; }
 
         #endregion IConnectionConfig Members
     }
